+++ plugins/sqlite/src/tests/oracle.test.ts
@@ -1,5 +1,5 @@
 import { describeQueryOracle } from '@routier/test-utils';
 import { uuidv4 } from '@routier/core';
-import { SqliteDbPlugin } from '../SqliteDbPlugin';
+import { SqliteDbPlugin } from '../index';
 
 describeQueryOracle('sqlite', () => new SqliteDbPlugin(`oracle-${uuidv4()}.sqlite`));