+++ plugins/sqlite/src/tests/products.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, afterAll } from '@jest/globals';
 import { generateData, seedData } from '@routier/test-utils';
 import { IDbPlugin, UnknownRecord, uuidv4 } from '@routier/core';
-import { SqliteDbPlugin } from '../SqliteDbPlugin';
+import { SqliteDbPlugin } from '../index';
 import { SqliteDataStore } from './data-access/context';
 import { faker } from '@faker-js/faker';